WildlifeCompliance: REFACTOR services removing fee policies.
